corrects file ctructure
